Crate egui_winit

egui bindings for winit.

The library translates winit events to egui, handled copy/paste, updates the cursor, open links clicked in egui, etc.

Feature flags

Re-exports

pub use egui;

pub use winit;

Modules

clipboard

window_settings 🔒

Macros

profile_function 🔒

Profiling macro for feature “puffin”

profile_scope 🔒

Profiling macro for feature “puffin”

Structs

EventResponse

State

Handles the integration between egui and winit.

WindowSettings

Can be used to store native window settings (position and size).

Functions

is_copy_command 🔒

is_cut_command 🔒

is_paste_command 🔒

is_printable_char 🔒

Winit sends special keys (backspace, delete, F1, …) as characters. Ignore those. We also ignore ‘\r’, ‘\n’, ‘\t’. Newlines are handled by the Key::Enter event.

native_pixels_per_point

open_url_in_browser 🔒

screen_size_in_pixels

translate_cursor 🔒

translate_mouse_button 🔒

translate_virtual_key_code 🔒