egui_winit

Function update_viewport_info

pub fn update_viewport_info(
    viewport_info: &mut ViewportInfo,
    egui_ctx: &Context,
    window: &Window,
    is_init: bool,
)

Update the given viewport info with the current state of the window.

Call before State::take_egui_input.

If this is called right after window creation, is_init should be true, otherwise false.