Crate egui_winit

egui bindings for winit.

The library translates winit events to egui, handled copy/paste, updates the cursor, open links clicked in egui, etc.

Feature flags

Re-exports

pub use accesskit_winit;

pub use egui;

pub use winit;

Modules

clipboard

Structs

EventResponse

State

Handles the integration between egui and a winit Window.

WindowSettings

Can be used to store native window settings (position and size).

Enums

ActionRequested

Functions

apply_viewport_builder_to_window

Applies what create_winit_window_builder couldn’t

create_window

Build and intitlaize a window.

create_winit_window_attributes

inner_rect_in_points

outer_rect_in_points

pixels_per_point

Calculate the pixels_per_point for a given window, given the current egui zoom factor

process_viewport_commands

screen_size_in_pixels

short_device_event_description

Short and fast description of a device event. Useful for logging and profiling.

short_window_event_description

Short and fast description of a window event. Useful for logging and profiling.

update_viewport_info

Update the given viewport info with the current state of the window.