Crate egui_winit

egui bindings for winit.

The library translates winit events to egui, handled copy/paste, updates the cursor, open links clicked in egui, etc.

Re-exports

pub use egui;

pub use winit;

Modules

clipboard

screen_reader

Structs

State

Handles the integration between egui and winit.

WindowSettings

Can be used to store native window settings (position and size).

Functions

native_pixels_per_point

screen_size_in_pixels