Struct egui_winit::State

pub struct State { /* private fields */ }

Handles the integration between egui and winit.

Implementations

impl State

pub fn new(max_texture_side: usize, window: &Window) -> Self

Initialize with:

max_texture_side: e.g. GL_MAX_TEXTURE_SIZE

the native pixels_per_point (dpi scaling).

pub fn from_pixels_per_point(
max_texture_side: usize,
pixels_per_point: f32
) -> Self

Initialize with:

max_texture_side: e.g. GL_MAX_TEXTURE_SIZE

the given pixels_per_point (dpi scaling).

pub fn pixels_per_point(&self) -> f32

The number of physical pixels per logical point, as configured on the current egui context (see egui::Context::pixels_per_point).

pub fn egui_input(&self) -> &RawInput

The current input state. This is changed by Self::on_event and cleared by Self::take_egui_input.

pub fn take_egui_input(&mut self, window: &Window) -> RawInput

Prepare for a new frame by extracting the accumulated input, as well as setting the time and screen rectangle.

pub fn on_event(&mut self, egui_ctx: &Context, event: &WindowEvent<'_>) -> bool

Call this when there is a new event.

The result can be found in Self::egui_input and be extracted with Self::take_egui_input.

Returns true if egui wants exclusive use of this event (e.g. a mouse click on an egui window, or entering text into a text field). For instance, if you use egui for a game, you want to first call this and only when this returns false pass on the events to your game.

Note that egui uses tab to move focus between elements, so this will always return true for tabs.

pub fn handle_platform_output(
&mut self,
window: &Window,
egui_ctx: &Context,
platform_output: PlatformOutput
)

Call with the output given by egui.

This will, if needed:

update the cursor

copy text to the clipboard

open any clicked urls

update the IME